Add notarize target to the makefile.
Added `notarize` target, with a chain of targets working backwards:

    notarize --> stapled.stamp.txt --> notarized.stamp.txt -->
        notarization-log.json --> submission-id.txt -->
        submit-log.json --> xz-$(version).pkg

Added the overridable makefile variable `NOTARIZATION_KEYCHAIN_PROFILE`
to point to the notarization credentials.

Added missing codesigning steps for the libraries `liblzma.a` and
`liblzma.5.dylib`.

Added additional steps to the `check` target to verify the libraries.

Updated the readme file with a "Prerequisites" section, info on the
`NOTARIZATION_KEYCHAIN_PROFILE` variable, a description of the
`notarize` target and a section on signing and notarization credentials.