Bump the package version to revision 2.
Prepare to change the build so that `xz` is statically linked to
`liblzma`.

Bump the package to rev 2.

Add `$(rev)` and `$(ver)` makefile variables.  `$(rev)` conditionally
builds the `-rX` string if rev is not `1`, and `$(ver)` incorporates the
revision.  Update targets to use `$(ver)` when specifying the package
name.

Add the `dylibs` wrapper around `otool` to the `tools` directory and add
tests to the `check` target to test that the executables don't have any
non-system dylib dependencies.

Update the readme with the new package name.