Fix bad dependency graph for the default target.
Update dependencies so that the install and package build steps are
sequenced correctly.

- `xz.pkg` now depends on `uninstall-xz` only
- `uninstall-xz` depends on `xz.path`
- `xz.path` depends on `installed-and-signed.stamp.txt`

Also change the action for `installed-and-signed.stamp.txt` so that it
doesn't delete the `install` directory each time it runs (this was
accidently removing `xz.path` and `uninstall.xz`.

The `uninstall-xz` target needs to happen after all other actions that
assemble the files to package, since its action walks the directory tree
to find all the files to remove.